Model completion and documentation of build process
Fix misc model errors
Fix some threading issues in python
Add documentation

+++ main.ts
@@ -2,16 +2,20 @@ import { app, BrowserWindow, screen } from 'electron'
 import * as path from 'path'
 import * as child from 'child_process'
 import * as fs from 'fs'
-import * as ping from 'ping'
 import * as url from 'url'
+import * as fetch from 'node-fetch'
+import * as pRetry from 'p-retry'
 
-const PY_DIST_FOLDER = 'pydist'
-const PY_MODULE = 'cli' // without .py suffix
-const SERVER_APP = url.fileURLToPath(`file://${__dirname}/${PY_DIST_FOLDER}/${PY_MODULE}.app`)
+const ENTRY_POINT = `file://${path.join(__dirname, 'dist/index.html')}`
+const PY_DIST_FOLDER = '../pydist'
+const PY_MODULE = 'gfunctionapibootstrapper' // without .py suffix
+
+const SERVER_APP = url.fileURLToPath('file://' + path.join(__dirname, PY_DIST_FOLDER) + '/' + PY_MODULE + '.app')
 const SERVER_ENDPOINT = 'http://localhost:1400/'
 const SERVER_PING = 'ping'
 const SERVER_SHUTDOWN = 'shutdown'
-let pyProc : child.ChildProcessWithoutNullStreams
+
+let pyProc  = null;
 let serverIsRunning = false
 
 let win: BrowserWindow = null
@@ -23,8 +27,8 @@ const serve = args.some(val => val === '--serve')
 function createWindow (): BrowserWindow {
   console.log('starting')
 
-  const electronScreen = screen
-  const size = electronScreen.getPrimaryDisplay().workAreaSize
+  const electronScreen = screen;
+  const size = electronScreen.getPrimaryDisplay().workAreaSize;
 
   // Create the browser window.
   win = new BrowserWindow({
@@ -34,26 +38,35 @@ function createWindow (): BrowserWindow {
     height: size.height,
     webPreferences: {
       nodeIntegration: true,
-      allowRunningInsecureContent: !!(serve),
-      contextIsolation: false, // false if you want to run 2e2 test with Spectron
-      enableRemoteModule: true // true if you want to run 2e2 test  with Spectron or use remote module in renderer context (ie. Angular)
-    }
-  })
+      allowRunningInsecureContent: (serve) ? true : false,
+      contextIsolation: false,  // false if you want to run 2e2 test with Spectron
+      enableRemoteModule : true // true if you want to run 2e2 test  with Spectron or use remote module in renderer context (ie. Angular)
+    },
+  });
 
-  console.log(win)
+  if (serve) {
+    console.log('serve')
+    win.webContents.openDevTools();
 
-  const entryPoint = url.fileURLToPath(`file://${__dirname}/dist/index.html`)
-  fileExists(entryPoint)
-  win.loadURL(entryPoint)
+    require('electron-reload')(__dirname, {
+      electron: require(`${__dirname}/node_modules/electron`)
+    });
+    win.loadURL('http://localhost:4200');
+
+  } else {
+    fileExists(ENTRY_POINT)
+    win.loadURL(url.format(new url.URL(ENTRY_POINT), {
+
+    }))
+  }
 
   // Emitted when the window is closed.
   win.on('closed', () => {
-    console.log('closing')
     // Dereference the window object, usually you would store window
     // in an array if your app supports multi windows, this is the time
     // when you should delete the corresponding element.
-    win = null
-  })
+    win = null;
+  });
 
   return win
 }
@@ -62,40 +75,41 @@ function createWindow (): BrowserWindow {
 
 const createPyProc = () => {
   fileExists(SERVER_APP)
-  // let scriptPath = path.resolve('cli.py');
-  // pyProc = child.execFile(script);
+  
   pyProc = child.spawn('open', [SERVER_APP])
   pyProc.addListener('error', function (err) {
     console.log(err)
   }).addListener('message', function (message) {
     console.log(message)
+  }).addListener('disconnect', function(d) {
+    console.log(d);
   })
 
-  pingServer()
+  return pingServer()
 }
 
 const pingServer = () => {
-  console.log(`ping server: ${SERVER_ENDPOINT}${SERVER_PING}`)
-  ping.sys.probe(`${SERVER_ENDPOINT}${SERVER_PING}`, pingResult, {
-    timeout: 5000
-  })
+    const pingResult = pRetry(fetchPage, {onFailedAttempt: error => {
+			console.log(`Attempt ${error.attemptNumber} failed. There are ${error.retriesLeft} retries left.`);
+		},
+		retries: 5})
+
+    return pingResult
 }
 
-const pingResult = (result) => {
-  console.log('result:' + JSON.stringify(result))
-  if (result) {
-    serverIsRunning = true
-  } else {
-    console.error('failed to load server')
+async function fetchPage () {
+  const response = await fetch(`${SERVER_ENDPOINT}${SERVER_PING}`, {})
+
+  // Abort retrying if the resource doesn't exist
+  if (response.status === 404) {
+    throw new pRetry.AbortError(response.statusText)
   }
+
+  return response.blob()
 }
 
 const shutdownServer = () => {
-  if (serverIsRunning) {
-    ping.sys.probe(`${SERVER_ENDPOINT}${SERVER_SHUTDOWN}`, shutDownResult, {})
-  }
-
-  serverIsRunning = false
+  fetch(`${SERVER_ENDPOINT}${SERVER_SHUTDOWN}`, {}).then(shutDownResult)
 }
 
 const shutDownResult = (result) => {
@@ -105,7 +119,7 @@ const shutDownResult = (result) => {
 const exitPyProc = () => {
   console.log('exiting py proc')
   shutdownServer()
-  pyProc.kill()
+  if (pyProc) pyProc.kill()
   pyProc = null
 }
 
@@ -122,8 +136,11 @@ const fileExists = (filePath: string) => {
 // Added 400 ms to fix the black background issue while using transparent window. More detais at https://github.com/electron/electron/issues/15947
 app.on('ready', () => {
   console.log('ready')
-  setTimeout(createWindow, 400)
-  setTimeout(createPyProc, 400)
+  const result = createPyProc()
+
+  result.then(() => setTimeout(createWindow, 400)).catch((err) => {
+    console.log(err);
+  })
 })
 
 // Quit when all windows are closed.
@@ -138,14 +155,14 @@ app.on('window-all-closed', () => {
 
 app.on('activate', () => {
   console.log('activating')
+  // if (pyProc === null) {
+  //   createPyProc()
+  // }
+
   // On OS X it's common to re-create a window in the app when the
   // dock icon is clicked and there are no other windows open.
   if (win === null) {
     createWindow()
-  }
-
-  if (pyProc === null) {
-    createPyProc()
   }
 })
